Update setup and configuration for puppeteer
