+++ server/classes/headstart/preprocessing/Snapshot.php
@@ -14,7 +14,7 @@ class Snapshot
         );
         $url_postfix = http_build_query($post_data);
 
-        $phantomjs = $ini_array["snapshot"]["phantomjs_path"];
+        $node_path = $ini_array["snapshot"]["node_path"];
         $getsvg = $ini_array["snapshot"]["getsvg_path"];
         if (isset($_SERVER) and array_key_exists('SERVER_PROTOCOL', $_SERVER)) {
           $protocol = stripos($_SERVER['SERVER_PROTOCOL'],'https') === true ? 'https://' : 'http://';
@@ -26,11 +26,10 @@ class Snapshot
         $host = $protocol . $ini_array["general"]["host"];
         $snap_php = $ini_array["snapshot"]["snapshot_php"];
         $storage = $ini_array["snapshot"]["storage_path"];
-        $width = $ini_array["snapshot"]["snapshot_width"];
         $nodemodules = $ini_array["snapshot"]["nodemodules_path"];
 
         $url = "{$host}{$snap_php}?{$url_postfix}";
-        $this->cmd = "{$phantomjs} {$getsvg} \"{$url}\" {$storage}{$post_data['file']}.png {$width} {$nodemodules}";
+        $this->cmd = "{$node_path} {$getsvg} \"{$url}\" {$storage}{$post_data['file']}.png {$nodemodules}";
     }
 
     public function takeSnapshot() {
