+++ server/preprocessing/conf/config.ini
@@ -7,7 +7,7 @@ preprocessing_dir = "/path/to/preprocessing/"
 # Full path to the images directory for the client. Needs to be in the public_html/www directory. Make sure that your webserver has write access to this directory.
 images_path = "/path/to/paper_preview/"
 # Host of the client visualization
-host = "http://localhost/"
+host = "localhost/"
 # Relative path to the client visualization. Needs to be in the public_html/www directory.
 vis_path = "path/to/vis"
 # Relative path to the client REST services. Needs to be in the public_html/www directory.
@@ -16,8 +16,8 @@ services_path = "server/services/"
 [snapshot]
 # Set to 1 to enable snapshot feature, 0 to disable
 snapshot_enabled = 0
-# Absolute path to phantomjs binary
-phantomjs_path = "/path/to/phantomjs"
+# Absolute path to node binary
+node_path = "/path/to/node"
 # Absolute paths to node_modules folder, where puppeteer is installed
 nodemodules_path = "/path/to/node_modules"
 # Absolute path to getChartSVG.js
@@ -26,8 +26,6 @@ getsvg_path = "/path/to/getChartSVG.js"
 storage_path = "/path/to/storage/"
 # PHP File responsible for rendering the bubble in a way to be snapshotted. Relative path to general host
 snapshot_php = "server/services/snapshot/headstart_snapshot.php"
-# Thumbnail width
-snapshot_width = "1200px"
 # snapshot_local_protocol fallback for non-server environments
 snapshot_local_protocol = "http://"
