+++ server/services/getChartSVG.js
@@ -1,73 +1,17 @@
-var page = require('webpage').create(),
-        system = require('system'),
-        address, output, size;
+module.paths.push(process.argv[4]);
+const puppeteer = require('puppeteer');
 
-if (system.args.length < 3 || system.args.length > 5) {
-    console.log('Usage: rasterize.js URL filename [paperwidth*paperheight|paperformat] [zoom]');
-    console.log('  paper (pdf output) examples: "5in*7.5in", "10cm*20cm", "A4", "Letter"');
-    console.log('  image (png/jpg output) examples: "1920px" entire page, window width 1920px');
-    console.log('                                   "800px*600px" window, clipped to 800x600');
-    phantom.exit(1);
-} else {
-    address = system.args[1];
-    output = system.args[2];
-    
-    page.viewportSize = {width: 600, height: 600};
-    if (system.args.length > 3 && system.args[2].substr(-4) === ".pdf") {
-        size = system.args[3].split('*');
-        page.paperSize = size.length === 2 ? {width: size[0], height: size[1], margin: '0px'}
-        : {format: system.args[3], orientation: 'portrait', margin: '1cm'};
-    } else if (system.args.length > 3 && system.args[3].substr(-2) === "px") {
-        size = system.args[3].split('*');
-        if (size.length === 2) {
-            pageWidth = parseInt(size[0], 10);
-            pageHeight = parseInt(size[1], 10);
-            page.viewportSize = {width: pageWidth, height: pageHeight};
-            page.clipRect = {top: 0, left: 0, width: pageWidth, height: pageHeight};
-        } else {
-            console.log("size:", system.args[3]);
-            pageWidth = parseInt(system.args[3], 10);
-            pageHeight = parseInt(pageWidth * 3 / 4, 10); // it's as good an assumption as any
-            console.log("pageHeight:", pageHeight);
-            page.viewportSize = {width: pageWidth, height: pageHeight};
-        }
-    }
-    if (system.args.length > 4) {
-        page.zoomFactor = system.args[4];
-    }
-    page.onError = function (msg, trace) {
+function timeout(ms) {
+    return new Promise(resolve => setTimeout(resolve, ms));
+};
 
-        var msgStack = ['ERROR: ' + msg];
+(async() => {
+    const browser = await puppeteer.launch();
+    const page = await browser.newPage();
+    await page.setViewport({width: 1920, height: 1080})
+    await page.goto(process.argv[2], {waitUntil: 'networkidle2'});
+    await timeout(2000)
+    await page.screenshot({path: process.argv[3], clip: { x: 0, y: 0, width: 1150, height: 1080 }});
+    browser.close();
 
-        if (trace && trace.length) {
-            msgStack.push('TRACE:');
-            trace.forEach(function (t) {
-                msgStack.push(' -> ' + t.file + ': ' + t.line + (t.function ? ' (in function "' + t.function + '")' : ''));
-            });
-        }
-
-        console.log(msgStack.join('\n'));
-
-    };
-
-    page.open(address, function (status) {
-        if (status !== 'success') {
-            console.log('Unable to load the address!');
-            phantom.exit(1);
-        } else {
-            window.setTimeout(function () {
-                page.viewportSize = { width: 1500, height: 1080 };
-                var svgClipRect = page.evaluate( function() {
-                    var visCol = document.getElementsByClassName("vis-col")[0];
-                    document.getElementById("infolink").remove();
-                    return visCol.getBoundingClientRect();
-                });
-                page.clipRect = svgClipRect;
-                window.setTimeout(function() {
-                    page.render(output);
-                    phantom.exit();
-                }, 1000);
-            }, 1000);
-        }
-    });
-}
+})();
