+++ server/services/snapshot/data-config_pubmed.js
@@ -16,4 +16,13 @@ var data_config = {
     area_force_alpha: 0.015,
     show_list: true,
     content_based: false,
+    
+    show_context: true,
+    create_title_from_context: true,
+    context_most_relevant_tooltip: true,
+    
+    doi_outlink: true,
+    filter_menu_dropdown: true,
+    sort_menu_dropdown: true,
+    filter_options: ["all", "open_access"],
 };
